Write an exploded stats TSV
Simple v1, for manual review of history stats. Does not evaluate.
Known issue: does not cleanly handle `git mv`

What to look for:
* If a file is changed in a large number of commits, those changes can
probably be squashed
* If a file has much more line churn than its overall diff, those
changes can probably be squashed
* If a single commit is very small, it might be squashable
* If a single commit touches a large number of files or lines, it might
be a candidate for splitting.
* If some the files changed in commits are exhibiting strong clustering,
this might be a candidate for splitting into multiple PRs.
* If a commit is pulling in cross-cluster files, it may be logically
conflating changes and may need to be split.